simplify extension
Bench: 2172304

+++ src/search.h
@@ -52,7 +52,7 @@ namespace NNUE {
 struct Networks;
 }
 
-inline constexpr std::uint8_t SEARCHED_MOVE_CAPACITY = 32;
+inline constexpr std::size_t SEARCHED_MOVE_CAPACITY = 32;
 
 using SearchedMoves = FixedVector<Move, SEARCHED_MOVE_CAPACITY, std::uint8_t>;
